Create functional user registratio
Created a modal with a user registration form that will be called on the ‘create account’ button on the login screen. The user data is saved in localStorage, under the ‘savedUsers’ key. The password recovery option will not be implemented, but when clicking on the ‘forgot your password?’ link, the user receives a message informing that the functionality is under development.