Create toolbar
Created a toolbar to be used on all pages, with the exception of the login screen. The button to enable and disable the sidebar-menu has been moved to the toolbar. The toolbar also displays the page title and the user’s name accompanied by a generic icon, since the user cannot upload an image in this version of the project. The loginService was also modified so that when authenticating the user, it now stores the user’s name instead of the email in the ‘loggedUser’ key in localStorage.