Create medical record list
Created a page that lists all registered patients. On this page, it is possible to filter patients by name. The list contains the patient’s identifier, name, and health insurance. An icon was also inserted next to the id, which will later serve as a button to open the details of each patient’s medical record.