Start method for editing patients
Created put method in DataService to edit JSON Server information.
Corrected exams and appointments ID logic to allow editing later.
Included edit button in the patient data area on the medical records page.
When clicking on the edit button, it takes you to the patient registration route,
passing the id as a parameter in the route and pulling the patient's data into the form.