Implement functional record and data editing
Finished editing option for exams and medical appointments. When clicking to edit a certain exam or appointment in the patient’s medical record, the id of the exam or appointment is passed to the route of their registration form. The buttons are enabled and functional. The date formats have been corrected to the DD-MM-YYYY format.