Create patients info component and allow editing and removal patients
Created a component to handle patient cards that were displayed on the home page. A field was also created to filter patients. The card contains basic patient information and a button that redirects to the patient registration page, passing the patient’s id as a parameter and displaying all the patient’s registration data in this form. It is now possible to edit the patient’s data. Enabled the option to delete the registration when there is no exam or consultation linked to the patient.