Correct date format in patient registration
Fixed date format saved in the JSON Server for date fields in the patient registration form. Also fixed a bug that did not allow saving changes to the patient registration when there was no health plan expiration date filled out. Since this field is not mandatory, it should fill an empty string in the database.