Fix validation for id and name fields in appointments and exams forms
Disabled id and name fields in the appointments and exams registration forms. Previously, it was possible to fill in any patient id and name, and the form would be saved, even without the patient having a record. Now the fields are disabled and can only be filled through the setPatientData method.