connect appointment form to backend
The patient search in the appointment registration route has been changed to render the search result just below the search input. When clicking on the desired patient, the patient's name and ID are filled in the respective fields of the form. The service and appointment registration component have been adjusted to connect to the database.