connect exam form to backend
The partient search in the exam registration route has been changed to render the search result just below the search input. When clicking on the desired patient, the patient's name and ID are filled in the respective fields of the form. The service and exam registration component have been adjusted to connect to the database.